feat: Rename file module to sourcing/filesystem
